% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.angles.R
\name{conventions}
\alias{conventions}
\title{Plot angle conventions}
\usage{
conventions(type = c("pitch", "yaw", "view_elevation"))
}
\arguments{
\item{type}{Character vector specifying which convention plot(s) to draw.
Choices are \code{"pitch"}, \code{"yaw"}, and \code{"view_elevation"}. Default plots all.}
}
\value{
No return value.
}
\description{
Quick visual reminders of the angle conventions used by \code{araponga}.
}
\examples{

conventions()

conventions("yaw")

}
\seealso{
\code{\link[=plot.angles]{plot.angles()}}
}
