% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\alias{rad2deg}
\title{Convert degrees to radians, and vice versa}
\usage{
deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{deg}{numeric; angle to be converted, in degrees}

\item{rad}{numeric; angle to be converted, in radians}
}
\value{
converted angle, in radians or degrees
}
\description{
Convert degrees to radians, and vice versa
}
\section{Functions}{
\itemize{
\item \code{deg2rad()}: Convert degrees to radians

\item \code{rad2deg()}: Convert radians to degrees

}}
\examples{
deg2rad(180)
rad2deg(pi)
}
