% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation2d.from.3d.R
\name{elevation2d.from.3d}
\alias{elevation2d.from.3d}
\title{Calculate projected 2D elevation angle from orientations and view inclination}
\usage{
elevation2d.from.3d(elevation, azimuth, view_inclination, plot = FALSE)
}
\arguments{
\item{elevation}{numeric \link{90, -90}; vertical angle relative to horizontal plane, in degrees (up = 90, down = -90)}

\item{azimuth}{numeric (-180, 180]; horizontal angle around the vertical axis, in degrees (0 = pointed right, 90 = pointed straight away, -90 = pointed straight torward, 180 = pointed left)}

\item{view_inclination}{numeric \link{-90, 90}; angle between observer and the object, in degrees (-90 straight below, 0 = eye level, 90 = straight up)}

\item{plot}{logical; if TRUE plots calculated 2D elevation angle.}
}
\value{
2D elevation angle, in degrees
}
\description{
Calculate projected 2D elevation angle from orientations and view inclination
}
\examples{
# object pointed up 15 degrees
elevation2d.from.3d(15, 0, 0, plot = TRUE)
# object pointed up 15 and right 30 degrees
elevation2d.from.3d(15, -30, 0, plot = TRUE)
# object pointed up 15 degrees, looked at from 30 degrees below
elevation2d.from.3d(15, 0, -30, plot = TRUE)
}
