% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation2d.from.xy.R
\name{elevation2d.from.xy}
\alias{elevation2d.from.xy}
\title{Calculate projected 2D elevation from two landmarks}
\usage{
elevation2d.from.xy(x_tip, y_tip, x_base, y_base, plot = FALSE)
}
\arguments{
\item{x_tip}{numeric; x coordinate of object tip.}

\item{y_tip}{numeric; y coordinate of object tip.}

\item{x_base}{numeric; x coordinate of object base.}

\item{y_base}{numeric; y coordinate of object base.}

\item{plot}{logical; if TRUE plots calculated 2D elevation angle.}
}
\value{
2D elevation angle, in degrees
}
\description{
Calculate projected 2D elevation from two landmarks
}
\examples{
# pointed right
elevation2d.from.xy(1,0,0,0, plot = TRUE)
#' # pointed left
elevation2d.from.xy(-1,0,0,0, plot = TRUE)
#' # pointed up
elevation2d.from.xy(0,1,0,0, plot = TRUE)
#' # pointed down
elevation2d.from.xy(0,-1,0,0, plot = TRUE)
}
