% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.azimuth.R
\name{find.azimuth}
\alias{find.azimuth}
\title{Find the azimuth that projects a known 2D elevation, given known elevation and view inclination}
\usage{
find.azimuth(elevation2d, elevation, view_inclination)
}
\arguments{
\item{view_inclination}{}
}
\description{
Find the azimuth that projects a known 2D elevation, given known elevation and view inclination
}
