% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.candidate.azimuths.R
\name{find.candidate.azimuths}
\alias{find.candidate.azimuths}
\title{Find azimuths that can project a known 2D elevation, given known view inclination and candidate elevations}
\usage{
find.candidate.azimuths(
  candidate.azimuths,
  candidate.elevations,
  elevation2d,
  view_inclination,
  plot = TRUE,
  branch = "both"
)
}
\arguments{
\item{candidate.azimuths}{}

\item{candidate.elevations}{}

\item{elevation2d}{}

\item{view_inclination}{}

\item{plot}{}

\item{branch}{}
}
\description{
Find azimuths that can project a known 2D elevation, given known view inclination and candidate elevations
}
