% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.candidate.elevations.R
\name{find.candidate.elevations}
\alias{find.candidate.elevations}
\title{Find elevations that can project a known 2D elevation, given known view inclination and candidate azimuths}
\usage{
find.candidate.elevations(
  candidate.elevations,
  candidate.azimuths,
  elevation2d,
  view_inclination,
  plot = TRUE,
  return = "elevation"
)
}
\arguments{
\item{elevation2d}{}

\item{plot}{}

\item{return}{}
}
\description{
Find elevations that can project a known 2D elevation, given known view inclination and candidate azimuths
}
