% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.elevation.R
\name{find.elevation}
\alias{find.elevation}
\title{Find the elevation that projects a known 2D elevation, given known azimuth and and view inclination}
\usage{
find.elevation(elevation2d, azimuth, view_inclination)
}
\arguments{
\item{view_inclination}{}
}
\value{
elevation elevation angle, in degrees
}
\description{
Find the elevation that projects a known 2D elevation, given known azimuth and and view inclination
}
