% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.pitch.range.R
\name{find.pitch.range}
\alias{find.pitch.range}
\title{Find the 3D pitch that projects a known 2D pitch, given known 3D roll and range of possible 3D yaws and pitches}
\usage{
find.pitch.range(
  pitch.range,
  yaw.range,
  pitch2d,
  roll3d,
  plot = TRUE,
  return = "pitch"
)
}
\arguments{
\item{pitch.range}{}

\item{yaw.range}{}

\item{pitch2d}{}

\item{roll3d}{}

\item{plot}{}

\item{return}{}
}
\description{
Find the 3D pitch that projects a known 2D pitch, given known 3D roll and range of possible 3D yaws and pitches
}
