% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.pitch3d.R
\name{find.pitch3d}
\alias{find.pitch3d}
\title{Find the 3D pitch that projects a known 2D pitch, given known 3D yaw and roll}
\usage{
find.pitch3d(pitch2d, yaw3d, roll3d)
}
\arguments{
\item{pitch2d}{projected 2D pitch angle, in degrees; e.g., from {pitch2d_from_xy}}

\item{yaw3d}{}

\item{roll3d}{}
}
\description{
Find the 3D pitch that projects a known 2D pitch, given known 3D yaw and roll
}
