% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.yaw.R
\name{find.yaw}
\alias{find.yaw}
\title{Find 3D yaws that project to a known 2D pitch (given 3D pitch and view elevation)}
\usage{
find.yaw(
  pitch2d,
  view_elevation,
  pitches = NULL,
  yaws = NULL,
  round2keep = 0,
  plot = FALSE
)
}
\arguments{
\item{pitch2d}{Numeric scalar: observed 2D pitch as returned by \code{\link[=pitch2d.from.xy]{pitch2d.from.xy()}}, in degrees in
the interval (-180, 180]}

\item{view_elevation}{Numeric scalar: angle between camera and object, in degrees, in the
interval [-90, 90]. Convention: \code{-90} = seen from straight below, \code{0} = eye level,
\code{90} = seen from straight above.}

\item{pitches}{Optional numeric vector of one known or multiple candidate pitch angles, in
degrees, in the interval [-90, 90]. Convention: \code{90} = pointed up, \code{0} = horizontally
aligned, \code{-90} = pointed down. If \code{NULL} (default), a grid comprehending all possible
pitches (\code{seq(-90, 90, 0.01)}) is used.}

\item{yaws}{Optional numeric vector of candidate 3D yaws used to filter the
results, in degrees, in the interval (-180, 180]. Convention: \code{0} = pointed right,
\code{90} = pointed straight away, \code{-90} = pointed straight toward, \code{180} = pointed left. If
\code{NULL} (default) all computed yaws are returned.}

\item{round2keep}{Integer scalar: number of decimal places to pass to \code{\link[=round]{round()}} when comparing
equality between computed and candidate pitches/yaws. Default is \code{0}, i.e., equality is
checked after rounding computed and candidate values to integers.}

\item{plot}{Logical scalar. If \code{TRUE}, the function plots the full continuous yaw-by-pitch
curves, overlays lines for the selected \code{yaws}/\code{pitches}, and points for the returned
combinations.}
}
\value{
A \code{data.frame} with two columns:
\describe{
\item{pitches}{numeric: the (input) pitch values, in degrees, retained after filtering}
\item{yaws}{numeric: the corresponding computed 3D yaws, in degrees (0 = pointed right,
90 = pointed straight away, -90 = pointed straight toward, 180 = pointed left).}
}
If no (pitch, yaw) matches are found the returned data.frame has zero rows.
}
\description{
Given a measured 2D pitch and one or more candidate pitch orientations, compute the
corresponding yaw angles that would produce that projected 2D pitch at a given view elevation.
}
\details{
The function \code{\link[=find.pitch]{find.pitch()}} in practice finds which solutions of \code{\link[=pitch2d.from.3d]{pitch2d.from.3d()}} reproduce \code{pitch2d}.

More specifically, \code{\link[=pitch2d.from.3d]{pitch2d.from.3d()}} constructs the full rotation matrix \eqn{R} using
\code{\link[=rotate3d]{rotate3d()}} and computes the projected 2D pitch \eqn{p_{2}} as
\deqn{p_{2} = \operatorname{atan2}\!\big(R_{2,1},\,R_{1,1}\big),}

where:
\deqn{R_{2,1} = \sin(p_{3})\cos(r_{3}) + \sin(r_{3})\sin(y_{3})\cos(p_{3}),}
\deqn{R_{1,1} = \cos(p_{3})\cos(y_{3}),}

\eqn{r_{3}} is \code{view_elevation}, \eqn{p_{3}} is a candidate/known 3D \code{pitch}, and \eqn{y_{3}}
is the yaw we are interested in.

We find \eqn{y_{3}} by inverting this relation:
\deqn{\phi=\operatorname{atan2}(-\sin(r_{3}),\,\tan(p_{2}))}
\deqn{y_{3} = \phi \pm \arccos\!\left(\dfrac{\cos(r_{3})\;\tan(p_{3})}{\sqrt{\tan(p_{2})^{2}+\sin(r_{3})^{2}}}\right)}

This expression is implemented inside \code{\link[=find.yaw]{find.yaw()}}.
}
\examples{
# yaws that project to 10° seen from 15° below
find.yaw(pitch2d = 10, view_elevation = -15, plot = TRUE)

# similar to above, but now given known pitch of 5°
find.yaw(pitch2d = 10, view_elevation = -15, pitches = 5, plot = TRUE)

# similar to above, but now given range of candidate pitches instead (0 to 35°)
find.yaw(pitch2d = 10, view_elevation = -15, pitches = seq(0, 30, 0.1), plot = TRUE)

# similar to above, but now restricting (integer rounding) to range of candidate yaws (0 to 15°)
find.yaw(pitch2d = 10, view_elevation = -15, pitches = seq(0, 30, 0.1),
           yaws = seq(0, 15, 0.1), plot = TRUE)

}
\seealso{
\code{\link[=pitch2d.from.xy]{pitch2d.from.xy()}}, \code{\link[=find.pitch]{find.pitch()}}, \code{\link[=pitch2d.from.3d]{pitch2d.from.3d()}}
}
