% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch2d_from_3d.R
\name{pitch2d_from_3d}
\alias{pitch2d_from_3d}
\title{Calculate projected 2D pitch angle from 3D Euler rotations}
\usage{
pitch2d_from_3d(pitch3d, yaw3d, roll3d, order = "ZYX", plot = FALSE)
}
\arguments{
\item{pitch3d}{numeric; rotation about the z-axis (pitch), in degrees.}

\item{yaw3d}{numeric; rotation about the y-axis (yaw), in degrees.}

\item{roll3d}{numeric; rotation about the x-axis (roll), in degrees}

\item{order}{character; order of rotations to be applied, a three-letter string composed of X, Y, Z. Default "ZYX" means rotate about Z by pitch3d then Y by yaw3d then X by roll3d.}

\item{plot}{logical; if TRUE plots calculated 2D pitch angle.}
}
\value{
2D pitch angle, in degrees
}
\description{
Calculate projected 2D pitch angle from 3D Euler rotations
}
\examples{
# object pointed up 15 degrees
pitch2d_from_3d(15, 0, 0, plot = TRUE)
# object pointed up 15 and right 30 degrees
pitch2d_from_3d(15, -30, 0, plot = TRUE)
# object pointed up 15 degrees, looked at from 30 degrees below
pitch2d_from_3d(15, 0, -30, plot = TRUE)
}
