% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch2d_from_xy.R
\name{pitch2d_from_xy}
\alias{pitch2d_from_xy}
\title{Calculate projected 2D pitch angle from two landmarks}
\usage{
pitch2d_from_xy(x_tip, y_tip, x_base, y_base, plot = FALSE)
}
\arguments{
\item{x_tip}{numeric; x coordinate of object tip.}

\item{y_tip}{numeric; y coordinate of object tip.}

\item{x_base}{numeric; x coordinate of object base.}

\item{y_base}{numeric; y coordinate of object base.}

\item{plot}{logical; if TRUE plots calculated 2D pitch angle.}
}
\value{
2D pitch angle, in degrees
}
\description{
Calculate projected 2D pitch angle from two landmarks
}
\examples{
# pointed right
pitch2d_from_xy(1,0,0,0, plot = TRUE)
#' # pointed left
pitch2d_from_xy(-1,0,0,0, plot = TRUE)
#' # pointed up
pitch2d_from_xy(0,1,0,0, plot = TRUE)
#' # pointed down
pitch2d_from_xy(0,-1,0,0, plot = TRUE)
}
