% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\alias{rad2deg}
\title{Convert degrees to radians, and vice versa}
\usage{
deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{deg}{Numeric vector of angles in degrees.}

\item{rad}{Numeric vector of angles in radians.}
}
\value{
A numeric vector of angles, in radians or degrees. Length matches \code{deg} or \code{rad}.
}
\description{
\code{deg2rad()} takes angles in degrees and returns them converted in radians.

\code{rad2deg()} takes angles in radians and returns them converted in degrees.
}
\examples{
deg2rad(c(0, 90, 180))

rad2deg(c(0, pi/2, pi))

}
