% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch2d.w.error.R
\name{pitch2d.w.error}
\alias{pitch2d.w.error}
\title{Simulate error in 2D pitch calculation from labeling error}
\usage{
pitch2d.w.error(pitch2d, label_error, label_nsim = 100)
}
\arguments{
\item{pitch2d}{Numeric scalar: 2D pitch as returned by \code{\link[=pitch2d.from.xy]{pitch2d.from.xy()}}, in degrees in
the interval (-180, 180].}

\item{label_error}{Positive numeric scalar specifying the error (± pixels) in landmark
labeling.}

\item{label_nsim}{Integer scalar (default 100). Number of Monte Carlo draws passed internally to
\code{\link[=runif]{runif()}} to simulate error in pitch2d calculation.}

\item{plot}{Logical scalar. \code{TRUE} plots original and expanded sets of angles side by side.}
}
\value{
\code{lookup.error()} returns a named list with two elements:
\describe{
\item{\code{pitch}}{a list of maximum errors in pitch per source of error (named).}
\item{\code{yaw}}{a list of maximum errors in yaw per source of error (named).}
}

\code{include.error()} returns a numeric vector with the expanded candidate angles (pitches or yaws)
after incorporating the maximum error retrieved by \code{lookup.error()$pitch} or \code{lookup.error()$yaw}
with a 0.01 resolution. The maximum error is stored as vector in the attribute \code{error.included}.
}
\description{
\code{pitch2d.w.error()} simulates the effect of labeling inaccuracy on \code{pitch2d}
calculation.
}
\seealso{
\code{\link[=find.pitch]{find.pitch()}}, \code{\link[=find.yaw]{find.yaw()}}, \code{\link[=pitch2d.from.xy]{pitch2d.from.xy()}}
}
