% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.angles.R
\name{plot.angles}
\alias{plot.angles}
\title{Visualize angles}
\usage{
\method{plot}{angles}(
  angles = NULL,
  type = NULL,
  facing = c("right", "left"),
  col = NULL,
  main = NULL,
  labels = TRUE,
  add = FALSE
)
}
\arguments{
\item{angles}{Set(s) of angles to be plotted, in degrees in the range (-180, 180]. If a numeric vector,
one plot is generated. If a \code{list} or a \code{data.frame}, \code{length(angles)} plots are generated,
customized when \code{names(angles)} match \code{"pitch"}, \code{"yaw"}, or \code{"view_elevation"}. Compatible with
outputs from \code{\link[=find.3d]{find.3d()}}, \code{\link[=find.pitch]{find.pitch()}}, and \code{\link[=find.yaw]{find.yaw()}}.}

\item{type}{When \code{angles} is a vector, optional character choice between \code{"pitch"}, \code{"yaw"}, and
\code{"view_elevation"} that determines the kind of visualization to be generated. If \code{NULL} (default),
a generic plot is generated.}

\item{facing}{Character choice between \code{"right"} (default) or \code{"left"}: which direction the object
should be facing for \code{pitch} and \code{view_elevation}.}

\item{col}{Optional vector of colors to draw angle segments in each plot. Recycled to match
\code{length(angles)} when \code{angles} is not a vector.}

\item{main}{Optional character vector: titles for each plot. Recycled to match \code{length(angles)} when
\code{angles} is not a vector.}

\item{labels}{Logical scalar (default \code{TRUE}): should inset labels be plotted?}

\item{add}{Logical scalar (default \code{FALSE}): should angle segments be added to the current plotting
device?}
}
\value{
No return value.
}
\description{
Plotting helper for quick visual inspection of angles, in particular pitches, yaws, and view
elevations.
}
\examples{
# calls with a vector
## pitch between 60 and 70
plot.angles(60:70, "pitch")
## view_elevation between -30 and -40
plot.angles(-30:-40, "view_elevation")

# call with named list
list <- list(pitch = 10:30, yaw = -100:-45, view_elevation = -10:-15)
plot.angles(list)

# call with output from find.3d
## pitches and yaws that project to 10° (± 1° error) if seen from 15° below
df <- find.3d(
  9:11,
  find = c("pitch", "yaw"),
  candidate_view_elevations = -15
  )
plot.angles(df)

}
\seealso{
\code{\link[=find.3d]{find.3d()}}
}
