% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate3d.R
\name{rotate3d}
\alias{rotate3d}
\alias{Rx}
\alias{Ry}
\alias{Rz}
\title{Create and apply 3D Euler rotations}
\usage{
Rx(roll3d)

Ry(yaw3d)

Rz(pitch3d)
}
\arguments{
\item{roll3d}{numeric; rotation about the x-axis, in degrees}

\item{yaw3d}{numeric; rotation about the y-axis, in degrees}

\item{pitch3d}{numeric; rotation about the z-axis, in degrees}
}
\value{
a 3x3 rotation matrix
}
\description{
Create 3x3 rotation matrices from Euler angles (roll, yaw, pitch).
}
\section{Functions}{
\itemize{
\item \code{Rx()}: Rotate about the x-axis

\item \code{Ry()}: Rotate about the y-axis

\item \code{Rz()}: Rotate about the z-axis

}}
\examples{
# rotate 30 degrees about x
Rx(30)
# rotate 45 about z, then 30 about y, then -80 about x
Rx(-80) \%*\% Ry(30) \%*\% Rz(45)
}
