% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate3d.R
\name{rotate3d}
\alias{rotate3d}
\alias{Rx}
\alias{Ry}
\alias{Rz}
\title{Create and apply 3D Euler rotations}
\usage{
rotate3d(pitch, yaw, roll)

Rx(roll)

Ry(yaw)

Rz(pitch)
}
\arguments{
\item{pitch}{numeric; rotation about the z-axis, in degrees}

\item{yaw}{numeric; rotation about the y-axis, in degrees}

\item{roll}{numeric; rotation about the x-axis, in degrees}
}
\value{
a 3x3 rotation matrix with Euler rotations applied
}
\description{
Create 3x3 rotation matrices after applying pitch-yaw-roll ("ZYX") Euler rotations.
}
\section{Functions}{
\itemize{
\item \code{rotate3d()}: Create apply Euler rotations (roll, yaw, pitch)

\item \code{Rx()}: Rotate about the x-axis

\item \code{Ry()}: Rotate about the y-axis

\item \code{Rz()}: Rotate about the z-axis

}}
\examples{
# rotate 30 degrees about x
Rx(30)
# rotate 45 about z, then 30 about y, then -80 about x
rotate3d(45, 30, -80)
}
