% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.yaws.R
\name{summarize.yaws}
\alias{summarize.yaws}
\title{Summarize yaw angles into the smallest continuous interval}
\usage{
summarize.yaws(yaws)
}
\arguments{
\item{yaws}{Numeric vector of yaw angles, in degrees, in the interval (-180, 180].}
}
\value{
A named \code{list} with components:
\describe{
\item{from}{numeric scalar: interval start (more counterclockwise endpoint), in degrees.}
\item{to}{numeric scalar: inverval end (more clockwise endpoint), in degrees.}
\item{width}{numeric scalar: angular width of the shortest enclosing arc, in degrees.}
\item{wrap}{logical: \code{TRUE} if the interval crosses the -180/180 boundary.}
}
}
\description{
Given a set of yaw angles, \code{summarize.yaws()} returns the endpoints and width of the smallest
continuous angular interval that contains all supplied yaws.
}
\examples{
# simple non-wrapping interval
summarize.yaws(-60:60)

# a wrapping interval (points around ±180)
summarize.yaws(c(-175:-170, 171:179))

# output from find.yaw()
yaws <- find.yaw(pitch2d = 15, view_elevation = -45, pitches = seq(10, 20, 0.1))$yaws
summarize.yaws(yaws)

# singleton
summarize.yaws(30)

}
