% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.yaws.R
\name{trim.yaws}
\alias{trim.yaws}
\title{Trim candidate yaw sets known or expected to differ in yaw}
\usage{
trim.yaws(ccw_yaws, cw_yaws, min_sep, max_sep, plot = FALSE)
}
\arguments{
\item{ccw_yaws}{Numeric vector: candidate yaw angles known/expected to be
\strong{counterclockwise} of \code{cw_yaws} (degrees, in the interval (-180, 180]).}

\item{cw_yaws}{Numeric vector: candidate yaw angles known/expected to be \strong{clockwise} of
\code{ccw_yaws} (degrees, in the interval (-180, 180]).}

\item{min_sep}{Non-negative numeric scalar: the minimum angular separation (degrees) allowed
between a value in one set and a partner in the other set.}

\item{max_sep}{Non-negative numeric scalar: the maximum angular separation (degrees) allowed
between a value in one set and a partner in the other set.}

\item{plot}{Logical scalar. \code{TRUE} draws diagnostic plots with retained (green) and excluded
(red) angles for each set.}
}
\value{
A \code{list} with elements:
\describe{
\item{trimmed_ccw_yaws}{numeric: subset of \code{ccw_yaws} that have at least one matching \code{cw_yaws}.}
\item{trimmed_cw_yaws}{numeric: subset of \code{cw_yaws} that have at least one matching \code{ccw_yaws}.}
}
}
\description{
Given two sets of candidate yaw angles, \code{\link[=trim.yaws]{trim.yaws()}} removes elements from each set that
do not have at least one partner in the other set that is at least \code{min_sep} and up to
\code{max_sep} clockwise (for \code{ccw_yaws} set) or counterclockwise (for \code{cw_yaws}) from it. In
other words, each retained \code{ccw_yaws} angle will be ≥ \code{min_sep} and ≤ \code{max_sep}
counterclockwise of at least one \code{cw_yaws} angle. Conversely, each retained \code{cw_yaws} angle
will be ≥ \code{min_sep} and ≤ \code{max_sep} clockwise of at least one \code{ccw_yaws} angle.
}
\examples{
# hypothetical candidate yaw sets:
a <- 10:80
b <- 30:90

# we know `a` to be CCW of `b` by up to 180°
trim.yaws(a, b, 0, 180, plot = TRUE)

# we know `a` to be CCW of `b` by 30 to 45°
trim.yaws(a, b, 30, 45, plot = TRUE)

# if no mutual partners exist, both returned vectors become empty
trim.yaws(ccw_yaws = c(-10), cw_yaws = c(130), min_sep = 0, max_sep = 180)

}
